refactor: extract platform service helpers
